docs (manifest): semver compliance (with version reset)
realized too late after modifying the version numbers that i should've looked into semver. i just did and the previous numbers can't be mistaken for semver, and no one else ever used the extension, so i figured i go ahead with a reset.

decided to go with 1.0.0 because the extension is in a good enough state that i'm going to submit it to the web store in a few minutes, and this sounds the first major/public version to me ("public" excluding it being in a public github repo this whole time... because i'm extremely confident it's been getting 0 attention from anyone but me :P)