refactor (settings.js): create settings wrapper to facilitate better code sequence
previously had to loop through the settings and append them after everything else was appended, and that was an eyesore. declaring a wrapper and appending all the settings (and the settings title) inside allows the loop to be moved up in the code to right after the settings elems are created, making the final portion of the code much more readable

(also removed some unnecessary line-breaks which i think ijust accidentally in a recent commit. 1+ anecdote in favor of adding Prettier to the project)